Create README.md
Employee Attrition is a very important metric for all organizations. Employee Attrition when an employee decides to leave the company through any channel like resignation, layoffs, or even failure to come back after illness, etc. If that position is not refilled by a company for a long time that is employee attrition. Predicting who will leave or stay is very cost-effective for employers so that they can implement some strategies to maintain.